refactor: テーマ入力画面Container/Presentationalパターン廃止 #175
